--- a/changes/1732-PrettyWood.md
+++ b/changes/1732-PrettyWood.md
@@ -1,0 +1,1 @@
+support `default_factory` with `BaseModel.construct`