--- a/changes/1732-PrettyWood.md
+++ b/changes/1732-PrettyWood.md
@@ -1,1 +1,3 @@
-support `default_factory` with `BaseModel.construct`
+Add `default_factory` support with `BaseModel.construct`.
+Please note that **`__field_defaults__` has been removed**.
+Use `.get_default()` method on fields in `__fields__` attribute instead.