<!-- Thank you for your contribution! -->
<!-- Unless your change is trivial, please create an issue to discuss the change before creating a PR -->
<!-- See https://pydantic-docs.helpmanual.io/contributing/ for help on Contributing -->

# Change Summary
Some changes have been made on `default_factory` to avoid calling it for nothing and avoid some side effects in #1504 and #1712. And [this line](https://github.com/samuelcolvin/pydantic/pull/1504/files#diff-1ef405a6b5a2fa9f20ea7e1f92e4c38eR299) changed the default value that is used in `BaseModel.construct` to `None`.
Instead of changing `__field_defaults__`, which would be a mess, let's just use `ModelField.get_default` in `construct`.
The feature is hence scoped to `BaseModel.construct` and has the exact same behaviour as `BaseModel.__init__`

## Related issue number
closes #1732

<!-- Are there any issues opened that will be resolved by merging this change? -->

## Checklist

* [x] Unit tests for the changes exist
* [x] Tests pass on CI and coverage remains at 100%
* [x] Documentation reflects the changes where applicable
* [x] `changes/<pull request or issue id>-<github username>.md` file added describing change
  (see [changes/README.md](https://github.com/samuelcolvin/pydantic/blob/master/changes/README.md) for details)