@PrettyWood 2020-07-27
Hello @zoopp 
To be honest I checked the `__field_defaults__` while doing the PR and I agree it could be removed.
Right now we don't support `dict(exclude_defaults)` with `default_factory` and I wanted to remove `__field_defaults__` while working on it. But maybe we could remove `__field_defaults__` in this PR by just calling `BaseField.default`