@zoopp 2020-07-27
Sure, I was only inquiring about it out of curiosity.