@PrettyWood 2020-10-10
There is an effect on performance for sure

##### with `default_factory`
```
# 3 loops on `master` branch with 10 ** 6 loops
5704059017ns
5680524864ns
5950163218ns

# 3 loops on this branch with 10 ** 6 loops
5811149041ns
6014411367ns
5877212718ns
```

#### without `default_factory` (was not supported)
```
# 3 loops on `master` branch with 10 ** 6 loops
1517463041ns
1480426577ns
1481139567ns

# 3 loops on this branch with 10 ** 6 loops
1648662578ns
1656794397ns
1632753169ns
```

Don't know if it's a big deal.
We could maybe change the approach by keeping  `__field_defaults__` but not having a regular `dict` but something like
```python
class FieldDefaultsDict(Dict[str, Any]):
    def add_field_default(self, key: str, field: ModelField) -> None:
        if field.default_factory:
            default_value = field.default_factory
            default_value.__default_factory__ = True  # this would need some hack as `default_value` can be a builtin
            self[key] = default_value
        else:
            self[key] = field.default

    def __getitem__(self, item: str) -> Any:
        default_value = super().__getitem__(item)
        return default_value() if default_value.__default_factory__ else default_value
```